// src/editorjs-smde.type.ts
export type MarkdownEditorToolBarConfig = (
  | MarkdownEditorToolBarName
  | '|'
  | MarkdownEditorToolBarItem
)[];

export interface MarkdownEditorToolBarItem {
  name: MarkdownEditorToolBarName | string;
  action?: ((editor: SimpleMarkdownEditor) => void) | string;
  className?: string;
  title?: string;
}
export type MarkdownEditorToolBarName =
  | 'bold'
  | 'italic'
  | 'strikethrough'
  | 'heading'
  | 'heading-smaller'
  | 'heading-bigger'
  | 'heading-1'
  | 'heading-2'
  | 'heading-3'
  | 'code'
  | 'quote'
  | 'unordered-list'
  | 'ordered-list'
  | 'clean-block'
  | 'link'
  | 'image'
  | 'choose-image'
  | 'table'
  | 'horizontal-rule'
  | 'preview'
  | 'side-by-side'
  | 'fullscreen'
  | 'undo'
  | 'redo'
  | 'guide';

export interface SimpleMarkdownEditor extends ToolBarFuncs {
  /**
   * codemirror's EditorFromTextArea variable，can do something like attach event, and editor staff.
   */
  codemirror: EditorFromTextArea;
  /**
   * get，set
   * get: use empty param，reutrn markdown string
   * set: use string param，return SimpleMDE instance
   */
  value: (value?: string) => SimpleMarkdownEditor | string;
  /**
   * check is preview
   */
  isPreviewActive: () => boolean;
  /**
   * check is side by side preview
   */
  isSideBySideActive: () => boolean;
  /**
   * check is fullscreen
   */
  isFullscreenActive: () => boolean;

  /**
   * get editor CodeMirrorstatus
   */
  getState: () => any;

  /**
   * You can revert to the initial textarea by calling the toTextArea method.
   * Note that this clears up the autosave (if enabled) associated with it. The textarea will retain any text from the destroyed SimpleMDE instance.
   */
  toTextArea: () => void;

  /**
   * codeMirror's method,  rarely used
   */
  _replaceSelection: (cm: any, active: any, startEnd: any, url: any) => void;
}
export type MarkdownEditorToolbarFuncName =
  | 'toggleBold'
  | 'toggleItalic'
  | 'toggleStrikethrough'
  | 'toggleBlockquote'
  | 'toggleHeadingSmaller'
  | 'toggleHeadingBigger'
  | 'toggleHeading1'
  | 'toggleHeading2'
  | 'toggleHeading3'
  | 'toggleCodeBlock'
  | 'toggleUnorderedList'
  | 'toggleOrderedList'
  | 'cleanBlock'
  | 'drawLink'
  | 'drawImage'
  | 'drawTable'
  | 'drawHorizontalRule'
  | 'undo'
  | 'redo'
  | 'togglePreview'
  | 'toggleSideBySide'
  | 'toggleFullScreen';

export type ToolBarFuncs = {
  [key in MarkdownEditorToolbarFuncName]: (
    editor: SimpleMarkdownEditor
  ) => void;
};